:truck: Move routes. - Move routes to same path root.
